% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{dijkstra:}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{A data frame representing the graph, with columns for edges and weights.}

\item{init_node}{The starting node for the algorithm.}
}
\value{
A vector of shortest distances from the starting node to all other nodes.
}
\description{
This function computes the shortest paths from a given starting node to all other nodes in a graph.
}
\references{
https://en.wikipedia.org/wiki/Dijkstra%27s_algorithm
}
