% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{euclidean:}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{The first number.}

\item{b}{The second number.}
}
\value{
The greatest common divisor of two numbers.
}
\description{
This function computes the greatest common divisor of two numbers.
}
\references{
https://en.wikipedia.org/wiki/Euclidean_algorithm
}
