% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikigraph.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{wiki_graph}
\format{
A data frame with the following columns:
\describe{
  \item{v1}{The edges of the graph}
  \item{v2}{The edges of the graph}
  \item{w}{The weight of the edge}
}
}
\source{
\url{https://en.wikipedia.org/wiki/Graph}
}
\usage{
wiki_graph
}
\description{
A dataset containing the graph data extracted from a Wikipedia page.
}
\references{
Wikipedia: Graph
}
\keyword{datasets}
